+++ documents-service/src/main/java/ru/alefilas/service/impls/ModerationServiceImpl.java
@@ -2,7 +2,9 @@ package ru.alefilas.service.impls;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.alefilas.dto.moderation.ModerationResult;
@@ -10,10 +12,15 @@ import ru.alefilas.dto.moderation.ModerationTicketDto;
 import ru.alefilas.model.document.Document;
 import ru.alefilas.model.moderation.ModerationStatus;
 import ru.alefilas.model.moderation.ModerationTicket;
+import ru.alefilas.model.permit.PermitType;
 import ru.alefilas.repository.ModerationRepository;
 import ru.alefilas.service.ModerationService;
+import ru.alefilas.service.access.AccessHelper;
 import ru.alefilas.service.exception.ModerationTicketNotFoundException;
 import ru.alefilas.service.mapper.ModerationMapper;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class ModerationServiceImpl implements ModerationService {
@@ -24,8 +31,16 @@ public class ModerationServiceImpl implements ModerationService {
     @Override
     @Transactional
     public Page<ModerationTicketDto> getAllTickets(int page) {
-        return moderationRepository.findAll(PageRequest.of(page, 10))
-                .map(ModerationMapper::modelToDto);
+        Pageable pageable = PageRequest.of(page, 10);
+
+        List<ModerationTicket> visibleTickets = moderationRepository.findAll()
+                .stream()
+                .filter(tic -> AccessHelper.checkAccessBoolean(tic.getDocument().getParentDirectory(), PermitType.MODERATE))
+                .collect(Collectors.toList());
+
+        Page<ModerationTicket> tickets = new PageImpl<>(visibleTickets, pageable, visibleTickets.size());
+
+        return tickets.map(ModerationMapper::modelToDto);
     }
 
     @Override
@@ -35,6 +50,8 @@ public class ModerationServiceImpl implements ModerationService {
         ModerationTicket ticket = moderationRepository
                 .findById(result.getTicketId())
                 .orElseThrow(() -> new ModerationTicketNotFoundException(result.getTicketId()));
+
+        AccessHelper.checkAccess(ticket.getDocument().getParentDirectory(), PermitType.MODERATE);
 
         if (result.getResult() != ModerationStatus.ON_MODERATION) {
             moderationRepository.deleteById(result.getTicketId());
